refactor: Car 객체를 domain 폴더의 하위로 이동
